+++ docs/sphinx/examples/cpp/dynamics/cross_resonance.cpp
@@ -12,7 +12,7 @@
 // ```
 
 #include "cudaq/algorithms/evolve.h"
-#include "cudaq/dynamics_integrators.h"
+#include "cudaq/algorithms/integrator.h"
 #include "cudaq/operators.h"
 #include "export_csv_helper.h"
 #include <cudaq.h>
