+++ docs/sphinx/targets/cpp/quera_basic.cpp
@@ -6,7 +6,7 @@
 // Assumes a valid set of credentials have been stored.
 
 #include "cudaq/algorithms/evolve.h"
-#include "cudaq/dynamics_integrators.h"
+#include "cudaq/algorithms/integrator.h"
 #include "cudaq/schedule.h"
 #include <cmath>
 #include <map>
