+++ unittests/dynamics/test_RungeKuttaIntegrator.cpp
@@ -8,7 +8,7 @@
 
 #include "CuDensityMatState.h"
 #include "CuDensityMatTimeStepper.h"
-#include "cudaq/dynamics_integrators.h"
+#include "cudaq/algorithms/integrator.h"
 #include "test_Mocks.h"
 #include <cmath>
 #include <gtest/gtest.h>
